\name{demTp63Gp1}
\alias{demTp63Gp1}
\title{gprege on TP63 expression time-series.}
\description{
  Demo script of Gaussian Process Regression and Estimation of Gene Expression on TP63 time-series data (see gprege.m).
  See Kalaitzis & Lawrence (2011) for a detailed discussion of the ranking algorithm and dataset used.
}
\usage{
  demTp63Gp1(fulldataset=FALSE)
}
\arguments{
  \item{fulldataset}{(Logical) TRUE downloads and uses the full dataset.}
}
\seealso{
  \code{ \link{gprege} }
}
\examples{
  demTp63Gp1(fulldataset=FALSE)
}

