\name{DGdat_p63}
\docType{data}
\alias{DGdat_p63_case1_GL}
\alias{DGdat_p63_case2_GL} 
\alias{DGdat_p63_case3_GL}
\title{
  BATS rankings (Angelini, 2007)
  Case 1: Delta error prior
  Case 2: Inverse Gamma error prior
  Case 3: Double Exponential error prior
}
\description{
  \describe{
    \item{DGdat_p63_case1_GL}{Case 1: Delta error prior}

    \item{DGdat_p63_case2_GL}{Case 2: Inverse Gamma error prior}

    \item{DGdat_p63_case3_GL}{Case 3: Double Exponential error prior}
  }
}
\usage{
  data(DGdat_p63)
}
\source{
  GEO: \url{http://www.ncbi.nlm.nih.gov/geo/},
  TSNI ranking: \url{genome.cshlp.org/content/suppl/2008/05/05/gr.073601.107.DC1/DellaGatta_SupTable1.xls}
}
\references{
  Della Gatta G, et al.
  Direct targets of the TRP63 transcription factor revealed by a combination of
  gene expression profiling and reverse engineering.
  \emph{Genome Research} 2008, 18(6):939.
}
\keyword{datasets}
